+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [0.8.3][2015-11-25]
+
+### Added
+
+* CLI option `-q/--quied`
+* Nested route `POST /posts/1/comments`
+
 ## [0.8.2][2015-10-15]
 
 ### Added